pyproject.toml
[project]
name = "eeg-preprocessing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi==2025.10.5",
    "charset-normalizer==3.4.4",
    "colorama==0.4.6",
    "contourpy==1.3.3",
    "cycler==0.12.1",
    "decorator==5.2.1",
    "fonttools==4.60.1",
    "idna==3.11",
    "jinja2==3.1.6",
    "joblib==1.5.2",
    "kiwisolver==1.4.9",
    "lazy-loader==0.4",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "mne==1.10.2",
    "numpy==2.3.4",
    "packaging==25.0",
    "pandas==2.3.3",
    "pillow==12.0.0",
    "platformdirs==4.5.0",
    "pooch==1.8.2",
    "pyparsing==3.2.5",
    "pyserial>=3.5",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "requests==2.32.5",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "seaborn==0.13.2",
    "six==1.17.0",
    "threadpoolctl==3.6.0",
    "tqdm==4.67.1",
    "tzdata==2025.2",
    "urllib3==2.5.0",
]
